+++ CountlyConnectionManager.m
@@ -506,6 +506,10 @@ static dispatch_once_t onceToken;
     }
 #endif
 
+    if([Countly.user hasUnsyncedChanges])
+    {
+        [Countly.user save];
+    }
 
     isSessionStarted = YES;
     lastSessionStartTime = NSDate.date.timeIntervalSince1970;
@@ -543,6 +547,11 @@ static dispatch_once_t onceToken;
     if (!isSessionStarted) {
         CLY_LOG_W(@"%s No session is running, this 'updateSession' will be ignored", __FUNCTION__);
         return;
+    }
+    
+    if([Countly.user hasUnsyncedChanges])
+    {
+        [Countly.user save];
     }
 
     NSString* queryString = [[self queryEssentials] stringByAppendingFormat:@"&%@=%d",