+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## XX.XX.XX
-* Mitigated an issue where non-queued requests were affected from request timeout settings, not anymore.
+* Mitigated an issue where non-queued requests were affected from request timeout settings.
 
 ## 25.4.6
 * Added the ability to record reserved events.