+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## XX.XX.XX
-* Updated resolution extraction to accommodate iOS 26 deprecations in UIScreen/UIScene APIs.
+* Updated resolution extraction to accommodate iOS 26 deprecations.
 
 ## 25.4.6
 * Added the ability to record reserved events.