+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## XX.XX.XX
+* Updated user properties caching mechanism according to sessions.
+
 ## 25.4.6
 * Added the ability to record reserved events.
 * Changed default log level from "CLYInternalLogLevelDebug" to "CLYInternalLogLevelVerbose".